% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_these}
\alias{assign_these}
\title{Assign named list to calling frame for setting up exports}
\usage{
assign_these(x)
}
\description{
Assign named list to calling frame for setting up exports
}
\keyword{internal}
