% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_impute.R
\name{make_impute}
\alias{make_impute}
\alias{make_impute_at}
\alias{make_impute_all}
\alias{make_impute_if}
\alias{make_imputes}
\title{Imputation metaprogramming}
\usage{
make_impute(fun)

make_impute_at(fun)

make_impute_all(fun)

make_impute_if(fun)

make_imputes(.na)
}
\arguments{
\item{fun}{value or function used for imputing. See Details.}
}
\description{
Create functions that use a function for imputing
}
\details{
These functions make mutate-style impute functions using \code{\link[=impute]{impute()}} and the
supplied \code{.na} argument.

\code{make_imputes} is a wrapper around the other functions and returns each as
a list.
}
\keyword{internal}
