% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_predict.R
\name{na_predict}
\alias{na_predict}
\title{na_predict}
\usage{
na_predict(x, object, data = x)
}
\arguments{
\item{x}{data}

\item{object}{object with predict method}

\item{data}{data object}
}
\description{
replace \code{NA} values by predictions of a model
}
