% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nacars}
\alias{nacars}
\alias{nacars_dt}
\alias{nairis}
\alias{nairis_dt}
\title{data with missing values}
\format{An object of class \code{data.frame} with 6 rows and 11 columns.}
\usage{
nacars
}
\description{
data with missing values
}
\details{
\strong{cars} and \strong{iris} data sets with missing data for demonstration purposes.
}
\keyword{datasets}
