% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ns_export}
\alias{ns_export}
\alias{ns_export_function_list}
\title{Export functions in list}
\usage{
ns_export(nms)

ns_export_function_list(x)
}
\arguments{
\item{nms}{character list of name to export}

\item{x}{list; named list of functions to export}
}
\description{
Export functions in list
}
\details{
This uses the \code{@evalNamespace} directive to declare exports for objects by
name or when provided in a function.

ns_export <- function(nms) {
sprintf("export(%s)", paste(nms, collapse = ","))
}

ns_export_named <-function(x) {
nms <- names(x)
sprintf("export(%s)", paste(nms, collapse = ","))
}

#' @evalNamespace ns_export(names(fns))
}
\seealso{
\itemize{
\item \code{\link[=assign_these]{assign_these()}}
}
}
\keyword{internal}
