% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_cols.R
\name{drop_cols_all_na}
\alias{drop_cols_all_na}
\alias{drop_cols_any_na}
\alias{drop_na_cols}
\title{Remove columns with missing values}
\usage{
drop_cols_all_na(.tbl)

drop_cols_any_na(.tbl)

drop_na_cols(.tbl)
}
\arguments{
\item{.tbl}{table-like object}
}
\value{
An object of the same class as \code{data} with cols containing all
\code{NA} values removed
}
\description{
Remove columns of a table whose values are all \code{NA} or who have any \code{NA}
}
\details{
\code{drop_cols_all_na} removes all cols whose only values are \code{NA}.
\code{drop_cols_any_na} removes columns that have any \code{NA}. They work on all
table-like objects.
}
\seealso{
\itemize{
\item \code{\link[dplyr:select]{dplyr::select()}}
}
}
